+++ packages/addons/_tests/_setup/suite.ts
@@ -35,6 +35,10 @@ export function setupTest<Addons extends AddonMap>(
 		kinds: Array<AddonTestCase<Addons>['kind']>;
 		filter?: (addonTestCase: AddonTestCase<Addons>) => boolean;
 		browser?: boolean;
+		preInstallAddon?: (o: {
+			addonTestCase: AddonTestCase<Addons>;
+			cwd: string;
+		}) => Promise<void> | void;
 	}
 ) {
 	const test = vitest.test.extend<Fixtures>({} as any);
@@ -85,13 +89,17 @@ export function setupTest<Addons extends AddonMap>(
 			})
 		);
 
-		for (const { variant, kind } of testCases) {
+		for (const addonTestCase of testCases) {
+			const { variant, kind } = addonTestCase;
 			const cwd = create({ testId: `${kind.type}-${variant}`, variant });
 
 			// test metadata
 			const metaPath = path.resolve(cwd, 'meta.json');
 			fs.writeFileSync(metaPath, JSON.stringify({ variant, kind }, null, '\t'), 'utf8');
 
+			if (options?.preInstallAddon) {
+				await options.preInstallAddon({ addonTestCase, cwd });
+			}
 			const { pnpmBuildDependencies } = await installAddon({
 				cwd,
 				addons,